proxy/requirements_cosmos.txt
# Cosmos DB Migration Requirements
# Install these for Cosmos DB support

# Core Flask (already installed)
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0

# Azure Cosmos DB SDK
azure-cosmos==4.5.1

# Keep existing dependencies
python-dateutil==2.8.2
requests==2.31.0
